+++ framework/contracts/account/manager/src/error.rs
@@ -105,4 +105,13 @@ pub enum ManagerError {
 
     #[error("Register of sub account failed")]
     SubAccountRegisterFailed {},
+
+    #[error("Can't renounce account, with active sub account")]
+    RenounceWithSubAccount {},
+
+    #[error("Can't propose Renounced governance, use update_ownership instead")]
+    ProposeRenounced {},
+
+    #[error("Can't create account with Renounced governance")]
+    InitRenounced {},
 }