+++ framework/contracts/account/manager/tests/proxy.rs
@@ -1,15 +1,16 @@
 use abstract_adapter::mock::{MockExecMsg, MockInitMsg};
 use abstract_integration_tests::{create_default_account, mock_modules, AResult};
 
-use abstract_core::manager::{ModuleInstallConfig, ModuleVersionsResponse};
+use abstract_core::manager::{InfoResponse, ModuleInstallConfig, ModuleVersionsResponse};
 use abstract_core::objects::fee::FixedFee;
+use abstract_core::objects::gov_type::GovernanceDetails;
 use abstract_core::objects::module::{ModuleInfo, ModuleVersion, Monetization};
 use abstract_core::objects::module_reference::ModuleReference;
 use abstract_core::objects::namespace::Namespace;
 use abstract_core::objects::{AccountId, ABSTRACT_ACCOUNT_ID};
 use abstract_core::version_control::UpdateModule;
 use abstract_core::{manager::ManagerModuleInfo, PROXY};
-use abstract_integration_tests::{init_mock_adapter, install_adapter};
+use abstract_integration_tests::*;
 use abstract_interface::*;
 use abstract_manager::contract::CONTRACT_VERSION;
 use abstract_manager::error::ManagerError;
@@ -355,5 +356,52 @@ fn install_multiple_modules() -> AResult {
     assert_eq!(s2_balance, vec![coin(42, "token1"), coin(500, "token2")]);
     take_storage_snapshot!(chain, "proxy_install_multiple_modules");
 
+    Ok(())
+}
+
+#[test]
+fn renounce_cleans_namespace() -> AResult {
+    let sender = Addr::unchecked(OWNER);
+    let chain = Mock::new(&sender);
+    let deployment = Abstract::deploy_on(chain.clone(), sender.to_string())?;
+
+    let account = deployment.account_factory.create_new_account(
+        AccountDetails {
+            name: "foo".to_string(),
+            description: None,
+            link: None,
+            namespace: Some("bar".to_owned()),
+            base_asset: None,
+            install_modules: vec![],
+        },
+        GovernanceDetails::Monarchy {
+            monarch: sender.to_string(),
+        },
+        None,
+    )?;
+
+    let namespace_result = deployment
+        .version_control
+        .namespace(Namespace::unchecked("bar"));
+    assert!(namespace_result.is_ok());
+
+    account
+        .manager
+        .update_ownership(cw_ownable::Action::RenounceOwnership)?;
+
+    let namespace_result = deployment
+        .version_control
+        .namespace(Namespace::unchecked("bar"));
+    assert!(namespace_result.is_err());
+
+    // Governance is in fact renounced
+    let acc_cfg: InfoResponse = account.manager.info()?;
+    assert_eq!(
+        acc_cfg.info.governance_details,
+        GovernanceDetails::Renounced {}
+    );
+
+    let account_owner = account.manager.ownership()?;
+    assert!(account_owner.owner.is_none());
     Ok(())
 }