+++ framework/packages/abstract-client/src/error.rs
@@ -22,4 +22,7 @@ pub enum AbstractClientError {
 
     #[error("Module not installed")]
     ModuleNotInstalled {},
+
+    #[error("Account is Renounced and does not have an owner.")]
+    RenouncedAccount {},
 }